requirements.txt
pandas
numpy
sqlalchemy
scikit-learn
requests
beautifulsoup4
python-dotenv
openai
apscheduler
matplotlib
betfairlightweight
betfairlightweight
